notificationService/requirements.txt
dnspython==2.6.1
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1
pymongo==4.7.3
python-dotenv==1.0.1
